requirement.txt
spotipy
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
rapidfuzz
PyJWT
requests